fix: replace deprecated long with int

+++ sentry_dingtalk/plugin.py
@@ -136,7 +136,7 @@ class DingtalkPlugin(CorePluginMixin, notify.NotificationPlugin):
         issue_link = group.get_absolute_url(params={"referrer": "dingtalk"})
 
         if signature:
-            timestamp = long(round(time.time() * 1000))
+            timestamp = int(round(time.time() * 1000))
             secret = signature
             secret_enc = bytes(secret).encode("utf-8")
             string_to_sign = "{}\n{}".format(timestamp, secret)
